requirements.txt
torch
torchvision
kaggle
scikit-learn
pandas
numpy
seaborn
matplotlib
streamlit
Pillow
opencv-python
streamlit-webrtc